Move titan and luna into individual files.
Now in the src/server/ directory
