+++ src/visitor/worker_visitor.hpp
@@ -1,13 +1,12 @@
 /*
- * Defines interfaces for work_queue visitors and network_queue visitors
- * (latter still TODO).
+ * Defines interfaces for worker visitors for Luna and Titan.
  */
 
 #ifndef SOFTWARE_QUEUE_VISITOR_HPP
 #define SOFTWARE_QUEUE_VISITOR_HPP
 
-#include "queue_items.hpp"
-#include "safe_queue.hpp"
+#include "../server/queue_items.hpp"
+#include "../server/safe_queue.hpp"
 #include "../adc/lib/adc_block.hpp"
 #include "../final/main_network_worker.hpp"
 
@@ -18,7 +17,7 @@
  * Based on a design by Arne Mertz.
  * See https://arne-mertz.de/2016/04/visitor-pattern-part-2-enum-based-visitor/
  */
-class work_queue_visitor {
+class worker_visitor {
     public:
         safe_queue<work_queue_item>& qw; // the work queue from the containing worker
         safe_queue<network_queue_item>& qn; // the network queue from the containing worker
@@ -43,14 +42,14 @@ class work_queue_visitor {
         /**
          * Operation corresponding to wq_none (an empty queue).
          */
-        virtual void visitNone(work_queue_item&) = 0;
+        void visitNone(work_queue_item&);
 
         /**
          * Operation for an unknown/unspecified action.
          */
-        virtual void visitDefault(work_queue_item&) = 0;
+        void visitDefault(work_queue_item&);
 
-        work_queue_visitor(safe_queue<work_queue_item>& qw,
+        worker_visitor(safe_queue<work_queue_item>& qw,
                            safe_queue<network_queue_item>& qn,
                            adc_block& adcs,
                            main_network_worker* nw_ref)
@@ -58,7 +57,7 @@ class work_queue_visitor {
         {
 
         }
-        ~work_queue_visitor() { }
+        ~worker_visitor() { }
 
         /**
          * Selects the correct task for the given work queue item and its queue,
@@ -85,72 +84,6 @@ class work_queue_visitor {
                 default: {
                     visitDefault(item);
                     break;
-                }
-            }
-        }
-};
-
-/**
- * An interface to "visit" an item in a network queue and perform the appropriate
- * task based on the item's action. (Like work_queue_visitor, just using different
- * actions.)
- *
- * NOTE: the visit methods in this interface have return values (bool), unlike those
- * in work_queue_visitor.
- *
- * Based on a design by Arne Mertz.
- * See https://arne-mertz.de/2016/04/visitor-pattern-part-2-enum-based-visitor/
- */
-class network_queue_visitor {
-    private:
-        /**
-         * Operation corresponding to nq_recv.
-         */
-        virtual bool visitRecv(network_queue_item&) = 0;
-
-        /**
-         * Operation corresponding to nq_send_ack.
-         */
-        virtual bool visitAck(network_queue_item&) = 0;
-
-        /**
-         * Operation corresponding to nq_send.
-         */
-        virtual bool visitSend(network_queue_item&) = 0;
-
-        /**
-         * Operation corresponding to nq_none.
-         */
-        virtual bool visitNone(network_queue_item&) = 0;
-
-        /**
-         * Operation corresponding to an unspecified/unknown action.
-         */
-        virtual bool visitDefault(network_queue_item&) = 0;
-
-    public:
-        network_queue_visitor() { }
-        ~network_queue_visitor() { }
-
-        /**
-         * Selects the correct task for the given network queue item and performs it.
-         */
-        bool visit(network_queue_item& item) {
-            switch(item.action) {
-                case nq_recv: {
-                    return visitRecv(item);
-                }
-                case nq_send_ack: {
-                    return visitAck(item);
-                }
-                case nq_send: {
-                    return visitSend(item);
-                }
-                case nq_none: {
-                    return visitNone(item);
-                }
-                default: {
-                    return visitDefault(item);
                 }
             }
         }
